// client/game/plugins/hd/hdObjectLightData.ts
export type HdLightType = "STATIC" | "FLICKER" | "PULSE";

export interface HdObjectLightDefinition { alignment: string; height: number; radius: number; strength: number; color: [number, number, number]; type: HdLightType; duration: number; range: number; fadeInDuration: number; }